Clarify where to go for Deployment notifications
A customer reached out to me saying they couldn't enter the page where they can add a webhook notification for deployments. It turned out to be a misunderstanding, where the customer tried to enter the Add integrations page (which is in fact not available anymore). I don't think this is a big issue, but being just a bit more descriptive should prevent anyone from misunderstanding this in the future I think.